+++ src/videodownlinkbandwidthpolicy/VideoPreference.ts
@@ -32,6 +32,10 @@ export default class VideoPreference implements Eq, PartialOrd {
     );
   }
 
+  clone(): VideoPreference {
+    return new VideoPreference(this.attendeeId, this.priority, this.targetSize);
+  }
+
   private static readonly LOW_BITRATE_KBPS = 300;
   private static readonly MID_BITRATE_KBPS = 600;
   private static readonly HIGH_BITRATE_KBPS = 1200;