+++ test/task/JoinAndReceiveIndexTask.test.ts
@@ -232,6 +232,32 @@ describe('JoinAndReceiveIndexTask', () => {
       expect(context.videoSubscriptionLimit).to.equal(defaultVideoSubscriptionLimit);
     });
 
+    it('should set the server-supports-compression value in the application context when server requests for compressed sdp', async () => {
+      const joinAckFrame = SdkJoinAckFrame.create();
+      joinAckFrame.wantsCompressedSdp = true;
+
+      const joinAckSignal = SdkSignalFrame.create();
+      joinAckSignal.type = SdkSignalFrame.Type.JOIN_ACK;
+      joinAckSignal.joinack = joinAckFrame;
+
+      const joinAckBuffer = SdkSignalFrame.encode(joinAckSignal).finish();
+      joinAckSignalBuffer = new Uint8Array(joinAckBuffer.length + 1);
+      joinAckSignalBuffer[0] = 0x5;
+      joinAckSignalBuffer.set(joinAckBuffer, 1);
+
+      await delay(behavior.asyncWaitMs + 10);
+      expect(signalingClient.ready()).to.equal(true);
+      new TimeoutScheduler(100).start(() => {
+        webSocketAdapter.send(joinAckSignalBuffer);
+      });
+      new TimeoutScheduler(200).start(() => {
+        webSocketAdapter.send(indexSignalBuffer);
+      });
+      await task.run();
+      expect(context.indexFrame).to.not.equal(null);
+      expect(context.serverSupportsCompression).to.equal(true);
+    });
+
     it('can run and send join with application metadata if valid', async () => {
       await delay(behavior.asyncWaitMs + 10);
       expect(signalingClient.ready()).to.equal(true);